+++ github3/api.py
@@ -34,7 +34,8 @@ def authorize(login, password, scopes, note='', note_url='', client_id='',
                         client_secret)
 
 
-def login(username=None, password=None, token=None, url=None):
+def login(username=None, password=None, token=None, url=None,
+          two_factor_callback=None):
     """Construct and return an authenticated GitHub session.
 
     This will return a GitHubEnterprise session if a url is provided.
@@ -43,6 +44,8 @@ def login(username=None, password=None, token=None, url=None):
     :param str password: password for the login
     :param str token: OAuth token
     :param str url: (optional), URL of a GitHub Enterprise instance
+    :param func two_factor_callback: (optional), function you implement to
+        provide the Two Factor Authentication code to GitHub when necessary
     :returns: :class:`GitHub <github3.github.GitHub>`
 
     """
@@ -50,7 +53,7 @@ def login(username=None, password=None, token=None, url=None):
 
     if (username and password) or token:
         g = GitHubEnterprise(url) if url is not None else GitHub()
-        g.login(username, password, token)
+        g.login(username, password, token, two_factor_callback)
 
     return g