+++ github3/github.py
@@ -857,9 +857,11 @@ class GitHub(GitHubCore):
               two_factor_callback=None):
         """Logs the user into GitHub for protected API calls.
 
-        :param str username: (optional)
-        :param str password: (optional)
-        :param str token: (optional)
+        :param str username: login name
+        :param str password: password for the login
+        :param str token: OAuth token
+        :param func two_factor_callback: (optional), function you implement to
+            provide the Two Factor Authentication code to GitHub when necessary
         """
         if username and password:
             self._session.basic_auth(username, password)
@@ -1250,7 +1252,7 @@ class GitHub(GitHubCore):
                 'Accept': 'application/vnd.github.v3.full.text-match+json'
                 }
 
-        url = self._build_url('search', 'repositories')
+        url = self._build_url('search', 'users')
         return SearchIterator(number, url, UserSearchResult, self, params,
                               etag, headers)